:tada: Add new modules and improve functionality
Several new scripts related to journal, decorators, core abstracts, and others have been added in this commit. The overall codebase has seen improvements in functionality. Significant upgrades include the development of a unique Singleton decorator, and the implementation of a "TimingDecorator" for performance analysis. Additionally, the decorators now have an advanced logging feature. Further, a set of new enumerate classes have been added to the Enums.py script in the constants directory. These collectively enhance the software robustness and functionality.