:art: Reorganized publish script and added Windows scripts for package building
The major change here is the reorganization of the python publish script, moving it from `deploy/` to `scripts/`. I've also added `build-package.ps1`, `upload-pypi.ps1`, and `upload-pypi-test.ps1` files into the new `scripts/upload/pypi/win/` directory. These are scripts for building the package and uploading it to PyPI from a Windows environment. In addition, `publish-pypi-script.cmd` was renamed to `publish-pypi-script.ps1` to use the more modern PowerShell instead of cmd.