:sparkles: Add orm module with general base model
Two new files have been added to the project under the 'orm' module. The `__init__.py` file initializes the module while the `general.py` file includes the implementation of a general base model for the ORM. The base model class includes methods for getting table and field properties and a save method.